check if config file exists before copying it

+++ broker-cli/scripts/install-config.sh
@@ -2,5 +2,11 @@
 echo "Making directory"
 DIRECTORY=~/.sparkswap
 mkdir -p $DIRECTORY
-echo "Copying sparkswap-sample.js to ~/.sparkswap/sparkswap.js"
-cp -n \"$(dirname $(which sparkswap))/../lib/node_modules/broker-cli/sample-sparkswap.js\" ~/.sparkswap/sparkswap.js
+
+if [ ! -f "$DIRECTORY/sparkswap.js" ]; then
+  echo "Copying sparkswap-sample.js to ~/.sparkswap/sparkswap.js"
+  cp -n \"$(dirname $(which sparkswap))/../lib/node_modules/broker-cli/sample-sparkswap.js\" ~/.sparkswap/sparkswap.js
+else
+  echo "Config file already exists, you can override it with your sample file by running:"
+  echo "cp -n \"$(dirname $(which sparkswap))/../lib/node_modules/broker-cli/sample-sparkswap.js\" ~/.sparkswap/sparkswap.js"
+fi
